ENH: Better hyperlinks to the bug tracker.
Still not perfect, but better. BUG:, bug, issue, and # all work as prefixes to a decimal integer number that represents a bug id in the associated bug tracker. One such reference is allowed in any given commit comment. If a reference is parsed (via regex matching) from the commit comment, then there will be a hyperlink to that bug in the bug tracker when displayed on the viewChanges or viewUpdate pages.

Significantly added to test_compressedtest.php to actually execute this code and exercise all the possible bug number formats... Prior to this, it was completely uncovered.

git-svn-id: https://www.kitware.com/svn/CDash/trunk@2790 91079597-3006-4846-9bd3-4f3fb3d139a7