BUG: Provide a processsubmissions.php-specific error handler function to clean up when a detectable php or MySQL error occurs while processing a submitted xml file.
This marks a submission as "failed with an error" (status = 4) when the error handler is called in an error situation. It also releases the lock on the submissionprocessor table so that further submissions will be processed immediately upon the next call to submit.php.

Also, for the case where multiple retry attempts are being made, after the ridiculously long stall timeout period we have set on cdash.org right now, add a new config setting CDASH_SUBMISSION_PROCESSING_MAX_ATTEMPTS which limits the number of times we will attempt to process an xml file before giving up. The default value for this is 5. After that many attempts, the submission is marked as "failed after too many retries" (status = 5), and processing continues with the next available queued submission.

These changes prevent the silly "spinning server doing nothing" behavior that has caused a few days worth of temporarily empty dashboards while waiting for humans to intervene.

Remember: you have to check whether there was an error inside the error handler, because php registered shutdown functions get called at normal script exit, too, even when there was no error.


git-svn-id: https://www.kitware.com/svn/CDash/trunk@2887 91079597-3006-4846-9bd3-4f3fb3d139a7