ENH: Add support for URL upload
If a file having extension ".url" is uploaded, the associated record in
uploadfile table will have both the field 'isurl' set to True and the field
'filename' set to the content of the url file.

The "ViewFiles" view will display two tables:
 - a table listing the URLs associated with the build
 - a table listing the files associated with the build

The function "GetUploadedFiles" has been renamed to "GetUploadedFilesOrUrls"
in both upload and project models.

Test has also been updated to check if a link associated with the URL is properly
added to the view.

Possible improvement:
 - Instead of adding a field 'isurl', a table uploadurls could be created

git-svn-id: https://www.kitware.com/svn/CDash/trunk@3002 91079597-3006-4846-9bd3-4f3fb3d139a7