BUG: 14363: Restore behaviour of time_difference() function
Thanks to Igor for the quick bug report, and the patch to fix it.

Restore behaviour of time_difference() function regarding positive times as
it was until r3338

Since r3338 build time duration (available in advanced view) is
often shown as 'just now' in my CDash. I find this confusing.

Also, correct typo of "millisecond"

git-svn-id: https://www.kitware.com/svn/CDash/trunk@3348 91079597-3006-4846-9bd3-4f3fb3d139a7