EHN: Add third party library group for subprojects.
Modify the subproject visualization to support multiple groups.
Modify manageSubproject UI to use dropdown instead of button to display
subproject groups, and a modify button to change group.

Right now, the group is hardcoded the same way as core/non-core, we might
want to migrate to use a subproject group sql table instead.